+++ tests/Integration/Momento.Sdk.Tests/Cache/CacheControlTest.cs
@@ -3,7 +3,7 @@ using System.Linq;
 using System.Threading.Tasks;
 using Momento.Sdk.Auth;
 
-namespace Momento.Sdk.Tests.Integration.Cache.Control;
+namespace Momento.Sdk.Tests.Integration.Cache;
 
 [Collection("CacheClient")]
 public class CacheControlTest : TestBase
