+++ tests/Integration/Momento.Sdk.Tests/Cache/ListTest.cs
@@ -6,7 +6,7 @@ using Momento.Sdk.Requests;
 using Momento.Sdk.Responses;
 using Momento.Sdk.Tests;
 
-namespace Momento.Sdk.Tests.Integration.Cache.Data;
+namespace Momento.Sdk.Tests.Integration.Cache;
 
 [Collection("CacheClient")]
 public class ListTest : TestBase
